feat: maintain the original PURL from ingested docs
Fixes #1440

The PURL spec (and external libs implementing it) is in too much flux
wrt encoding to rely on our ability to reconstruct PURL's after
storing only their components in the database. So we're now storing
the original PURL from the ingested SBOM along with its components.

NOTE: We store PURL's as a result of ingesting advisories, but not all
of them contain PURL's and some result in many PURL's being created
for each version range the advisory might affect. The only "original"
PURL's we store are from SBOM's and even those may be slightly
different from what's in the doc, as we are liberal with data we
accept but strict about rendering it as a valid PURL.

We've removed the get_purl proc from the DB as well.