feat: delete raw docs from storage when advisory/sbom is deleted
Fixes: #1936

If the commit of the database transaction deleting either an advisory
or an sbom succeeds, we attempt to delete the raw doc from storage. If
that fails, an error will be logged and returned to the user.

The risk of leaving an orphan doc in storage was deemed acceptable
relative to the risk of leaving the database in an inconsistent state
by deleting the doc from storage prior to commiting the transaction.