The road to affirming logs is paved with incompetence and dispair
Both the tracing-test and testing_logger crates attempt to set a
global variable that is only supposed to be set once and is already
set by other tests. Unless the test using them is the only one
running, failure occurs.

So instead of logging in the delete_doc fn, we return an appropriate
error message, if any, that is expected to be logged at warn level by
the caller.